🚚 Refactor(characters): Rename npc sections components
