+++ frontend/src/components/modules/characters/modals/panels/characteristics/NpcCharacteristic.tsx
@@ -1,10 +1,10 @@
 import INpc from "@/models/npc/INpc";
-import NpcAbilityScores from "@/components/modules/characters/modals/panels/characteristics/sections/npc/NpcAbilityScores";
+import NpcAbilityScores from "@/components/modules/characters/modals/panels/characteristics/sections/npc/NpcAbilityScoresSection";
 import NpcBattle from "@/components/modules/characters/modals/panels/characteristics/sections/npc/NpcFightingSection";
-import Profile from "@/components/modules/characters/modals/panels/characteristics/sections/npc/Profile";
-import StandardActions from "@/components/modules/characters/modals/panels/characteristics/sections/npc/StandardActions";
-import LegendaryActions from "@/components/modules/characters/modals/panels/characteristics/sections/npc/LegendaryActions";
-import LairActions from "@/components/modules/characters/modals/panels/characteristics/sections/npc/LairActions";
+import Profile from "@/components/modules/characters/modals/panels/characteristics/sections/npc/NpcProfileSection";
+import StandardActions from "@/components/modules/characters/modals/panels/characteristics/sections/npc/NpcStandardActionsSection";
+import LegendaryActions from "@/components/modules/characters/modals/panels/characteristics/sections/npc/NpcLegendaryActionsSection";
+import LairActions from "@/components/modules/characters/modals/panels/characteristics/sections/npc/NpcLairActionsSection";
 
 interface Props {
     npc : INpc;
